+++ src/lib/icons.js
@@ -1,10 +1,13 @@
+import { FaCode, FaUserGraduate, FaSquare, FaCircle } from "react-icons/fa";
+import { BsDiamondFill } from "react-icons/bs";
 import { MdTimer } from "react-icons/md";
 import { RiSlideshowFill } from "react-icons/ri";
 import { GrWorkshop } from "react-icons/gr";
 import { GoInfo } from "react-icons/go";
 import { ImLab } from "react-icons/im";
-import { FaCode, FaUserGraduate } from "react-icons/fa";
 import styled from "styled-components";
+
+import {} from "react-icons/fa";
 
 export const TopicIcon = ({ type, ...props }) =>
   type === "sample" ? (
@@ -55,3 +58,41 @@ export const Exercise = ({ color = "#898989", size = 25, ...props }) => (
 export const CourseLab = ({ color = "#898989", size = 25, ...props }) => (
   <FaUserGraduate size={size} color={color} {...props} />
 );
+
+const Double = styled.div`
+  float: left;
+  svg:last-of-type {
+    position: relative;
+    left: -5px;
+  }
+`;
+
+export const GreenCircle = ({ size = 25, ...props }) => (
+  <FaCircle color="limegreen" size={size} {...props} />
+);
+
+export const BlueSquare = ({ size = 25, ...props }) => (
+  <FaSquare color="#6c6cfd" size={size} {...props} />
+);
+
+export const BlackDiamond = ({ size = 25, ...props }) => (
+  <BsDiamondFill color="black" size={size} {...props} />
+);
+
+export const DoubleDiamond = ({ size = 25, ...props }) => (
+  <Double {...props}>
+    <BsDiamondFill color="black" size={size} />
+    <BsDiamondFill color="black" size={size} />
+  </Double>
+);
+
+export const Difficulty = ({ level, size = 12 }) =>
+  level === "intermediate" ? (
+    <BlueSquare size={size} />
+  ) : level === "advanced" ? (
+    <BlackDiamond size={size} />
+  ) : level === "expert" ? (
+    <DoubleDiamond size={size} />
+  ) : level === "beginner" ? (
+    <GreenCircle size={size} />
+  ) : null;
